Learnings from Storage on Browser
Since xml is ~120MB unpacked, it's not possible to use file API.

Approach chosen is to parse xml, save each record separated, and record
on IndexedDB locally.

This seems fast and reliable.

Next step is create a parser to do this, show options to the user and
start expanding for next items